Update: added test cases for category controllers
